REGRESSION (Safari 18.4): Declarative shadow DOM and custom elements do not work with parseUnsafeHTML https://bugs.webkit.org/show_bug.cgi?id=292968 rdar://151273154
Reviewed by Chris Dumez.

The regression was caused by the refactoring to support custom element registry per TreeScope.

When a shadow root is inserted into a document tree with a non-null custom-element registry,
we should update the associated custom element registry to that of the destination document
like elements except when the shadow root is specified to use a scoped custom element registry.

* LayoutTests/fast/custom-elements/connected-callback-expected.txt: Added.
* LayoutTests/fast/custom-elements/connected-callback.html: Added.
* Source/WebCore/dom/ShadowRoot.cpp:
(WebCore::ShadowRoot::insertedIntoAncestor):

Canonical link: https://commits.webkit.org/295050@main