css/css-view-transitions/pseudo-element-animations fails to read 'opacity' computed style. https://bugs.webkit.org/show_bug.cgi?id=293031 <rdar://151351674>
Reviewed by Antoine Quint.

The test is getting the computed style, and trying to read opacity, which should
work since that's being animated.

This is because computeRenderStyleForProperty detects that 'opacity' could be
accelerated, and tries to use the originating element's animated style (and then
the cached pseudo style off of it).

Change this code to actually lookup the keyframes and check for an accelerated
animation of the current property, and then access the animated style (of the
pseudo's renderer, not the originating element's) if so.

I am unsure why this test now dumps the test file into the output, but we pass
an extra check, and I'll followup to fix the remaining failure (which then makes
the weird output go away).

* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/pseudo-element-animations-expected.txt:
* Source/WebCore/css/ComputedStyleExtractor.cpp:
(WebCore::computeRenderStyleForProperty):
(WebCore::ComputedStyleExtractor::customPropertyValue const):
(WebCore::ComputedStyleExtractor::propertyValue const):
(WebCore::ComputedStyleExtractor::layerCount const):
* Source/WebCore/style/Styleable.cpp:
(WebCore::Styleable::isRunningAcceleratedTransformAnimation const):
(WebCore::Styleable::isRunningAcceleratedAnimationOfProperty const):
* Source/WebCore/style/Styleable.h:

Canonical link: https://commits.webkit.org/295141@main