[CMake] Enable -Wunsafe-buffer-usage-in-libc-call compiler warning https://bugs.webkit.org/show_bug.cgi?id=293223
Reviewed by Michael Catanzaro.

Add -Wunsafe-buffer-usage-in-libc-call to the different targets of the
CMake build. As with -Wunsafe-buffer-usage, the flag is added only for
the GTK, WPE, and JSCOnly ports when targeting an Unix-like system. In
particular JSCOnly when building for Windows will not get the compiler
flag added even if supported. No code changes are needed at the moment
to make the build pass with this patch, but it will still provide the
safeguarding going forward.

Canonical link: https://commits.webkit.org/295156@main