Remove CSSOMViewScrollingAPIEnabled preference https://bugs.webkit.org/show_bug.cgi?id=293294
Reviewed by Antti Koivisto.

Given that the iOS IMDB app is actively updated it will use an SDK that
is newer than the one from iOS 12. It seems almost impossible for end
users to still have an IMDB app from that era and also be able to
update to the latest iOS so this should be safe to remove.

Canonical link: https://commits.webkit.org/295198@main