[WPE] WPE Platform: click counter should be reset on mouse move when passing the threshold https://bugs.webkit.org/show_bug.cgi?id=293338
Reviewed by Alejandro G. Castro.

In layout tests it can happen that we press, move and release and then
we press again at the same point quickly. We shouldn't get a double
click in that case if we moved more than the double click distance. This
patch moves the threshold check to wpe_view_event in case we emitted a
muse move event and we were tracking the press count.

Canonical link: https://commits.webkit.org/295199@main