[Grid] Crash in copyUsedTrackSizesForSubgrid with subgrid and line name list. https://bugs.webkit.org/show_bug.cgi?id=293392 rdar://149687269
Reviewed by Matt Woodrow.

When a grid item becomes a subgrid with a line name list
(e.g. subgrid [a] [b] [c]) this can end up affecting the positioning of
the subgrid in the parent grid. This is because the line names become
the number of tracks the subgrid will span in the parent grid if it has
an automatic span based on the other grid placement properties.

In some scenarios, such as the testcase explained below, this can cause us to
use a stale position from the subgrid because we fail to perform item
placement in the parent with the new span for the subgrid. To fix this
we need to treat this case as any other change to a grid item's
placement property as that is what it basically is. We do this in
RenderBox::styleDidChange -> RenderBox::updateGridPositionAfterStyleChange.

* LayoutTests/fast/css-grid-layout/subgrid-with-line-name-list-crash-expected.txt: Added.
* LayoutTests/fast/css-grid-layout/subgrid-with-line-name-list-crash.html: Added.
After the initial layout:
1.) Remove a child from the subgrid
- This causes us to setNeedsItemPlacement on the parent, which is a
  normal grid at this point.
2.) Change aforementioned grid to a subgrid.
- Since we already setNeedsItemPlacement on the now subgrid we avoid
  setting the bit again which also avoids setNeedsItemPlacement on the
  parent.
3.) Compute the intrinsic logical widths of the parent grid because of
min-width: max-content.
- This ends up performing item placement on the subgrid but the
  placement on the parent during this step uses a temporary grid
  separate from the one being used during layout.
4.) Perform layout of the parent grid. This skips item placement (it was
never set) and causes us to use the old position of the subgrid.

Canonical link: https://commits.webkit.org/295262@main