[macOS] Color input components do not update to match their parent's used `appearance` https://bugs.webkit.org/show_bug.cgi?id=293381 rdar://148625484
Reviewed by Antti Koivisto, Wenson Hsieh, and Abrar Rahman Protyasha.

`appearance` is an implementation detail of the elements in the color input's
shadow tree. When the used appearance of color inputs is `none` (either
explicitly specified using `appearance`, or after devolution due to
border/background styling), the inner components should also have a used
`appearance` of `none`, to avoid mixing and matching native and primitive styles.

The behavior described above is already implemented by checking the parent style
during style adjustment. However, if the used `appearance` changes later on,
updates to the parent are not always reflected in the children.

Fix this by checking for used `appearance` in
`RenderStyle::descendantAffectingNonInheritedPropertiesEqual`. This ensures that
descendent styles will be recomputed if the parent's `appearance` changes,
including the children's `appearance.

* LayoutTests/fast/forms/color/color-input-appearance-none-inherited-expected.html: Added.
* LayoutTests/fast/forms/color/color-input-appearance-none-inherited.html: Added.
* LayoutTests/fast/forms/color/color-input-devolved-appearance-inherited-expected.html: Added.
* LayoutTests/fast/forms/color/color-input-devolved-appearance-inherited.html: Added.
* LayoutTests/fast/forms/color/color-input-swatch-wrapper-appearance-none-inherited-expected.html: Added.
* LayoutTests/fast/forms/color/color-input-swatch-wrapper-appearance-none-inherited.html: Added.
* Source/WebCore/rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::descendantAffectingNonInheritedPropertiesEqual const):

Restructure the logic to be like other `RenderStyle` comparisons.

Add the check for used `appearance`, as explained above.

Canonical link: https://commits.webkit.org/295282@main