Don't generate Const V128 values in BBQ after a guaranteed throw https://bugs.webkit.org/show_bug.cgi?id=293300 rdar://151570328
Reviewed by Yusuke Suzuki and David Degazio.

When generating a struct or array get on a null reference, BBQ currently
generates a value with a null value. This leads to a release assertion
failure during emitMoveConst, due to BBQ not expecting a Const V128 to
be generated. We should replace these with `topValue` to explicitly say
that this value is undefined and should not be used.

* JSTests/wasm/stress/wasmgc-const-v128.js: Added.
(hostref):
(eq_ref):
(module):
(instance):
* Source/JavaScriptCore/wasm/WasmBBQJIT64.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::addArrayGet):
(JSC::Wasm::BBQJITImpl::BBQJIT::addStructGet):

Canonical link: https://commits.webkit.org/295289@main