[MSE] Seeking in MP4 containing b-frame ends may displays frame out of order. https://bugs.webkit.org/show_bug.cgi?id=283564 rdar://140415210
Reviewed by Jer Noble and Youenn Fablet.

Make sure that any frames following the seek point that has an earlier presentation time
are marked as not to be displayed.
Testing is problematic as you need to pause to accurately performs a reftest.
The out of order frames are getting displayed extremely briefly to be
immediately replaced by the next decoded frame.

* Source/WebCore/platform/graphics/TrackBuffer.cpp:
(WebCore::TrackBuffer::reenqueueMediaForTime):

Canonical link: https://commits.webkit.org/295304@main