[macOS] Web content process may crash when context menu is invoked https://bugs.webkit.org/show_bug.cgi?id=293342 rdar://151711558
Reviewed by Abrar Rahman Protyasha.

Prior to these changes, if the context menu was invoked and the hit node
was in the shadow DOM, was not an element, and had no parent node, the web
content process would crash. This is due to an error in `prepareContextForQRCode`,
where we would attempt to downcast the node into an element. If this failed, we
attempted to get the node's parent element instead. In the scenario listed above,
the node is not an element and had no parent, so when we dereferenced the
non-existent parent, a crash occurred.

These changes resolve this issue by checking the lineage of
`parentOrShadowHostElement` instead of `protectedParentElement`, and ensures
it is non-null before dereferencing.

* Source/WebCore/page/ContextMenuController.cpp:
(WebCore::prepareContextForQRCode):
* Tools/TestWebKitAPI/Tests/mac/ContextMenuTests.mm:
(TestWebKitAPI::TEST(ContextMenuTests, ContextMenuTopLevelTextNodeInShadowDOMShouldNotCrash)):

Canonical link: https://commits.webkit.org/295316@main