Model only scales on XY on zoom scale change https://bugs.webkit.org/show_bug.cgi?id=293024 rdar://150719171
Reviewed by Tim Horton.

This PR addresses an issue where, upon changing the zoom scale (using the 'a' and 'A'
buttons in the context menu in the URL bar) the model only scales on XY but not Z. The
fix is that we need to apply a transform3D to the container view such that the zoom scale
also gets applied to the model along the layer's Z axis. This also addresses zooming
using a double tap and 2 handed magnification gesture.

* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.h:
* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.mm:
(-[WKWebView scrollViewDidZoom:]):
(-[WKWebView _setWebViewTransform3DForModel:]):

Canonical link: https://commits.webkit.org/295317@main