[Refactor] Make AudioBus::create() return a Ref https://bugs.webkit.org/show_bug.cgi?id=293395 rdar://problem/151814836
Reviewed by Jean-Yves Avenard.

`AudioBus::create()` currently returns a RefPtr because of assertion that the `numberOfChannels`
parameter is not greater than a constant maximum value. However, that maximum is only necessary
because a stack variable in `copyWithGainFrom()` is created with that maximum channel size.
That stack variable is unnecessary if the looping algorithm is slightly modified.

Whats worse, a number of callers of `AudioBus::create()` convert the RefPtr to a Ref via
`releaseNonNull()` without checking whether the return value is null, which is a potential
null dereference waiting to happen.

Remove the concept of `MaxBusChannels`, and change the return type of `AudioBus::create()`
to a Ref.

* Source/WebCore/Modules/webaudio/AudioParam.cpp:
(WebCore::AudioParam::AudioParam):
* Source/WebCore/Modules/webaudio/ConvolverNode.cpp:
(WebCore::ConvolverNode::setBufferForBindings):
* Source/WebCore/platform/audio/AudioBus.cpp:
(WebCore::AudioBus::create):
(WebCore::AudioBus::copyFromRange):
(WebCore::AudioBus::copyWithGainFrom):
* Source/WebCore/platform/audio/AudioBus.h:
* Source/WebCore/platform/audio/AudioDestinationResampler.cpp:
(WebCore::AudioDestinationResampler::AudioDestinationResampler):
* Source/WebCore/platform/audio/SharedAudioDestination.cpp:
* Source/WebCore/platform/audio/gstreamer/AudioFileReaderGStreamer.cpp:
(WebCore::AudioFileReader::createBus):

Canonical link: https://commits.webkit.org/295322@main