Underruns in WebContent Process audio thread lead to extra latency rdar://151701056 https://bugs.webkit.org/show_bug.cgi?id=293296
Reviewed by Jean-Yves Avenard.

If a stall or hang occurs on the audio thread in the WebContent process, it may not produce
enough samples for reading by the GPU process, leading to an underrun when rendering Web Audio.
When this occurs, the GPU process will render silence until data exists at point in the
ring buffer where the stall occurred. So, if the WebContent process hangs for 100ms and then
recovers, the GPU process will continue reading from the point in the ring buffer where the
stall began, and will render 100ms in the past upon recovery, adding 100ms of latency to the
output.

Instead, the GPU process should just advance the ring buffer read head when an overflow occurs.
The WebContent process _should_ eventually recover if the hang was temporary. If the Web Audio
graph is too complicated to process in the timeslice available, it may fall further and further
behind, but the behavior will be that no audio is ever output, which is arguably a better
user experience than having 128 samples randomly render surrounded by long periods of silence.

* Source/WebKit/GPUProcess/media/RemoteAudioDestinationManager.cpp:

Canonical link: https://commits.webkit.org/295323@main