[GTK][WPE] Simplify the web view creation in GLib API tests https://bugs.webkit.org/show_bug.cgi?id=293480
Reviewed by Adrian Perez de Castro.

Remove all createWebView static functions and add a non-static one that
receives construct parameters and automatically sets the web-context and
network-session when not provided. In the case of WPE it always sets the
backend too. This function returns a GRefPtr so we no longer need the
adoptView() function. This simplifies the web view creation in tests and
will easily allow to add support for running API tests with new WPE API.

Canonical link: https://commits.webkit.org/295332@main