Don't emit cfi_{start,end}proc directives on ARMv7 https://bugs.webkit.org/show_bug.cgi?id=293479
Reviewed by Justin Michaud.

Emitting the directives in the cpp file means that there are now
additional assembler directives (emitted by the compiler) and one of
those appears to function as an implicit .cfi_endproc as far as GNU as
is concerned. This results in a mismatched directives error and a
failing build, so disable this on ARMv7 for now.

Canonical link: https://commits.webkit.org/295338@main