Document::cssParserContext() should return a reference https://bugs.webkit.org/show_bug.cgi?id=291761 rdar://149561463
Reviewed by Sam Weinig.

This removes 500K string refcounts from Speedometer, primarily by enabling
InlineCSSStyleProperties::cssParserContext() to reuse Document::cssParserContext()
in the common case.

That's 500K / 34MM, so this patch is not a measurable speedup on its own; but
the reduction is countable, and this is also an opportunity to land
OptionalOrReference<T> with a real use case, to complement ValueOrReference<T>.

* Source/WTF/wtf/OptionalOrReference.h: Copied from Source/WTF/wtf/ValueOrReference.h.
(WTF::OptionalOrReference::OptionalOrReference): New helper type that's just
like ValueOrReference<T> but doesn't require T() to be efficient.

Canonical link: https://commits.webkit.org/295371@main